Move upstream bearer resolution into src/auth
Verbatim move of the proxy-JWT-to-upstream-token resolution (error
class, refresh coalescing, proactive refresh) out of the HTTP transport
module. The logic is auth-domain: it touches only the proxy config,
proxy store, JWT verification, and the OAuth flow helpers.
